Move components in shared folder
