+++ assets/components/new-contribution/ContributionAmount.jsx
@@ -7,10 +7,10 @@ import { connect } from 'react-redux';
 
 import { amounts, type Amount, type Contrib } from 'helpers/contributions';
 import { type CountryGroupId } from 'helpers/internationalisation/countryGroup';
+import { type CountryMetaData } from 'helpers/internationalisation/contributions';
 
 import SvgDollar from 'components/svgs/dollar';
 
-import { type CountryMetaData } from '../contributionsLandingMetadata';
 
 // ----- Types ----- //
